Wave grid experiment rewritten into wave search
The 'wave grid' experiment no longer exists, it has now become wave search.
Wave search uses 7 waves as inputs and needs both the cDAQ and the NI 6216 devices.
For the nidaq an addition IO script is written that is purely meant for the wave search script. It uses the  NI 6216 as input and NI cDAQ 9171 with NI 9264 as output
Synchronization of the IO is not done in a clean way: writing starts before reading and therefore read has a delay. To synchronize write and read, 50ms of zeros is added to the input data and an additional input channel is used with a spike after 50ms of input data. The IO script automatically detects where this happens in the output data and cuts off from 0 up to this point im the output data.

Co-Authored-By: mark-boon <mark-boon@users.noreply.github.com>