Debugged wave search, first principles test of SPSA
Wave search with loading data seems to be working now. To do: check whether the data is lagging behind after each load.
SPSA is tested for the first time and at least runs the algorithm. To do: make sure the loss function is correct and the plot the figures properly

Co-Authored-By: mark-boon <mark-boon@users.noreply.github.com>